Crate intel_mkl_tool

Helper crate of build.rs in intel-mkl-src crate.

This crate is responsible for finding static or dynamic library of Intel MKL installed in user system.

Structs

Config

Configuration for Intel MKL, e.g. mkl-static-lp64-seq

Library

Found MKL library

Enums

DataModel

Data model of library

LinkType

How to link MKL

Threading

How to manage thread(s)

Constants

OPENMP_RUNTIME_LIB

STATIC_EXTENSION

Lacked definition of std::env::consts

VALID_CONFIGS

Functions

mkl_dyn_libs

MKL Libraries to be loaded dynamically

mkl_file_name

Filename convention for MKL libraries.

mkl_libs

MKL Libraries to be linked explicitly, not include OpenMP runtime (iomp5)

openmp_runtime_file_name

Filename convention for OpenMP runtime.